Improve commandfile discovery for extensions installed via Composer. (#156)
When managing plugins / extensions with Composer, and placing them via Composer Installers, there will
be additional directories placed by Composer that we need to account for when finding commandfiles.

For example, Drush has the following recommended installer path:

    "drush/Commands/contrib/{$name}": ["type:drupal-drush"]

If the Composer project has a name similar to `my-cool-plugin`, then the old discovery code would
produce a namespace that included dashes, and that included the `contrib` portion of the path. Both
of these things are undesirable. PHP namespaces cannot contain dashes, and if the namespace in
the plugin contains "contrib", then it would not be possible to commit a copy of the repository
in a "custom" directory, or directly in "drush/Commands".

This PR transliterates '-' to '_' in namespaces, and also allows clients to define namespace
path components (e.g. 'custom' and 'contrib') that should be removed before returning the discovered
paths and namespaces to the client.